+++ src/nj.cpp
@@ -137,7 +137,7 @@ void NJOI::TREEMAKE (std::vector<std::string>& names, std::map<int, std::string>
     std::vector< std::vector<double> >& Matrix) {
 
     int mini1 = 0, mini2 = 0;
-    int NumbOfSequences = NumbKeys.size();
+    int NumbOfSequences = static_cast<int>(NumbKeys.size());
     double brlength1 = 0.0;
     double brlength2 = 0.0;
     std::vector< std::vector<double> > LengthMatrix(NumbOfSequences, std::vector<double>(NumbOfSequences, 0.0));
@@ -169,7 +169,7 @@ std::vector< std::vector<double> > NJOI::BuildMatrix (std::map<std::string, std:
     double MatchScore;
 
     // an easier way to initialize a std::vector of std::vectors:
-    int ntax = sequences.size();
+    int ntax = static_cast<int>(sequences.size());
     std::vector< std::vector<double> > Score(ntax, std::vector<double>(ntax, 0.0));
 
     //compare all sequences to other sequences
